refactor(gsplat): move GPU sort/cull pipeline into GSplatHybridRenderer
Step 5: the hybrid renderer now owns its radix sorter, projector, interval
compaction and indirect slots, and the cull/project/sort methods
(prepareRenderView, sortAndProjectForCamera, _runFrustumCulling,
computeDistanceRange, allocateAndWriteIntervalIndirectArgs). Picking moves to
the renderer's preparePickingView. The manager keeps the world, the version
lifecycle (markSorted/onSorted), and the CPU worker sorter, passing per-call
param bags to the renderer (no back-reference). Behavior-preserving.
